Rename `events` module to `updates`
