+++ identity/src/lib.rs
@@ -86,7 +86,7 @@ pub mod account {
   pub use identity_account::account::*;
   pub use identity_account::crypto::*;
   pub use identity_account::error::*;
-  pub use identity_account::events::*;
+  pub use identity_account::updates::*;
   pub use identity_account::identity::*;
   pub use identity_account::storage::*;
   pub use identity_account::stronghold::*;
